ACPI: battery: Work around Dell's spurious 0% readings
My Precision 5540 occasionally reports 0% battery capacity despite the
battery being well over 50% charge. A reboot after this occurs shows the
correct battery capacity, at which point the laptop can be used on battery
power for a few hours without issue, indicating that the battery really
wasn't at 0%.

The bug actually turns out to be that ACPI is reporting spurious 1% battery
readings that get interpreted as 0% due to precision loss. On Windows, this
precision loss doesn't occur, so many XPS 15 7590 owners have posted online
about a "false 1%" battery issue [1], which is the same bug.

Attempting to reread the battery state from ACPI doesn't fix the issue. So
far, rebooting or plugging in the charger seem to be immediate remedies.
Cycling through a suspend and then resume might fix it as well, or waiting
for a really long time could work. It also seems like this bug only ever
occurs when the laptop has been suspended at least once since boot, as it
only ever happens shortly after resume.

Inspecting the DSDT shows that there's unfortunately no alternative ACPI
method to read the battery state, so we're stuck with this issue.

In any case, just report the last good battery reading when this happens so
that the battery meter exhibits some sanity when this occurs, and doesn't
cause userspace to invoke its various critical battery level actions.

[1] https://www.dell.com/community/XPS/XPS-15-7590-false-battery-warning/td-p/7703662

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>